feat(frontend): Move League and Team Components from Routes to Features
- CreateLeague moved to Features/League
- JoinLeague moved to Features/League
- LeaguesHome renamed to Leagues, endpoint changed to leagues
- League renamed to SelectedLeague

fixes #186
