+++ frontend/src/Features/League/CreateLeague.jsx
@@ -1,5 +1,5 @@
 import { useState } from "react"
-import { createLeague } from "../api/League/backend_calls";
+import { createLeague } from "../../api/League/backend_calls";
 import { toast } from "react-toastify";
 import { useNavigate } from "react-router-dom"
